waterlevel/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0

numpy==1.26.4
pandas==2.1.4
scikit-learn==1.3.2
scipy==1.11.4
joblib==1.3.2

requests==2.31.0
mysql-connector-python==8.3.0

edge-tts==6.1.9
